Moved course material functions to library directory from model directory
