+++ site/app/views/course/CourseMaterialsView.php
@@ -3,8 +3,8 @@
 namespace app\views\course;
 
 use app\libraries\Core;
+use app\libraries\CourseMaterialsUtils;
 use app\libraries\DateUtils;
-use app\models\CourseMaterial;
 use app\models\User;
 use app\views\AbstractView;
 use app\libraries\FileUtils;
@@ -68,7 +68,7 @@ class CourseMaterialsView extends AbstractView {
                         }
 
                         // Determine file authorization based on this file 'user_allow_list'
-                        $authorized_by_allow_list[$expected_file_path] = CourseMaterial::isUserAllowedByAllowList($this->core->getUser()->getId(), $json, $expected_file_path);
+                        $authorized_by_allow_list[$expected_file_path] = CourseMaterialsUtils::isUserAllowedByAllowList($this->core->getUser()->getId(), $json, $expected_file_path);
 
                         if (isset($json[$expected_file_path]['external_link']) && $json[$expected_file_path]['external_link'] === true) {
                             $contents = json_decode(file_get_contents($expected_file_path));
